CHANGE _ structure project
