+++ src/components/Main/Main.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 // import SideBar from './SideBar/SideBar';
-import RouteSite from './NavigationSite/NavigationSite';
+import NavigationSite from './NavigationSite/NavigationSite';
 
 import css from './Main.module.css';
 
@@ -8,7 +8,7 @@ const Main = () => (
   <>
     <div className={css.wrapMain}>
       {/* <SideBar /> */}
-      <RouteSite />
+      <NavigationSite />
     </div>
   </>
 );
