+++ src/components/Main/NavigationSite/NavigationSite.jsx
@@ -1,17 +1,17 @@
 import React from 'react';
 import { Route, Switch } from 'react-router-dom';
 
-import Tasks from '../../../pages/Tasks/Tasks';
+import TasksPage from '../../../pages/TasksPage';
 import Contacts from '../SideBar/Contacts/Contacts';
-import WeatherPage from '../../../pages/Weather';
-import RatePage from '../../../pages/Rate';
-import ErrorPage from '../../../pages/Error';
-import LoginPage from '../../../pages/Login';
+import WeatherPage from '../../../pages/WeatherPage';
+import RatePage from '../../../pages/RatePage';
+import ErrorPage from '../../../pages/ErrorPage';
+import LoginPage from '../../../pages/LoginPage';
 
 const RouteSite = () => (
   <div>
     <Switch>
-      <Route path="/" exact component={Tasks} />
+      <Route path="/" exact component={TasksPage} />
       <Route path="/login" exact component={LoginPage} />
       <Route path="/contacts" component={Contacts} />
       <Route path="/weather" component={WeatherPage} />
