+++ src/components/Main/Tasks/Column/Column.module.css
@@ -11,7 +11,7 @@
 .title {
   padding: 8px;
 }
-.tasks{
-    padding: 8px;
-    flex-grow: 1;
+.tasks {
+  padding: 8px;
+  flex-grow: 1;
 }
