+++ src/components/Main/Tasks/Tasks.module.css
@@ -5,3 +5,7 @@
   position: relative;
   height: 87vh;
 }
+.container {
+  display: flex;
+  justify-content: space-between;
+}
