--- src/pages/Login.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const Login = () => <>Login Page</>;
-
-export default Login;
