+++ src/pages/LoginPage.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const LoginPage = () => <>Login Page</>;
+
+export default LoginPage;
