+++ src/pages/RatePage.js
