--- src/pages/Tasks/EnterDay/EnterDay.module.css
@@ -1,73 +1,0 @@
-.wrapEnterDay {
-  display: flex;
-  justify-content: space-between;
-}
-.wrapDataPicker {
-  display: flex;
-}
-.wrapIcon {
-  width: 30px;
-  height: 30px;
-  padding: 5px 5px 5px 10px;
-}
-.calendar {
-  color: grey;
-}
-.dataPicker {
-  height: 4vh;
-  padding: 10px 5px 5px 5px;
-  color: grey;
-  border: none;
-  font-weight: 700;
-  width: 110px;
-  outline: none;
-}
-
-.wrapTasksLevels {
-  justify-content: center;
-  align-items: center;
-  font-size: 20px;
-  display: flex;
-}
-/* .wrapAllTaskIcon {
-  border: 2px solid grey;
-  border-radius: 50%;
-  width: 20px;
-  height: 20px;
-  position: relative;
-  margin-right: 5px;
-}
-.allTaskIcon {
-  position: absolute;
-  color: grey;
-  margin-right: 5px;
-  bottom: -2px;
-  left: 5px;
-} */
-.allTaskBtn {
-  margin-right: 5px;
-  font-size: 14px;
-  border-radius: 10px;
-  background-color: transparent;
-  border: 1px solid grey;
-  outline: none;
-  box-shadow: 0 2px 3px #575555 inset;
-}
-.allTaskBtn:focus,
-.allTaskBtn:hover {
-  background-color: grey;
-  color: white;
-}
-
-.red {
-  color: #ff7675;
-  margin-right: 5px;
-}
-
-.yellow {
-  color: #ffeb3b;
-  margin-right: 5px;
-}
-.green {
-  color: #a9f374;
-}
