--- src/pages/Tasks/Tasks.module.css
@@ -1,11 +1,0 @@
-.wrapTasks {
-  padding: 5px;
-  margin: 0 auto;
-  border: 1px solid #bdc3c7;
-  position: relative;
-  height: 87vh;
-}
-.container {
-  display: flex;
-  justify-content: space-between;
-}
