+++ src/pages/TasksPage.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import Tasks from '../components/Main/Tasks/Tasks';
+
+const TasksPage = () => (
+  <>
+    <Tasks />
+  </>
+);
+
+export default TasksPage;
